refactor: move function to utils folder
